refactor: amountToHangul 상수 분리
